Many security fixex and types added

+++ application/views/gallery.php
@@ -3,16 +3,29 @@
 <div id="index">
     <?php if ($_GET['list']=='grid'): ?>
         <?php foreach ($files as $file): ?>
-            <div class="demo-card-image mdl-card mdl-shadow--2dp item" data-type="image" data-src="/files/render/o/<?=$file['path']['name'];?>" style="background: url('/files/render/s/<?=$file['path']['name'];?>') center / cover">
-                <div class="mdl-card__title mdl-card--expand"></div>
-                <div class="mdl-card__actions">
-                    <span class="demo-card-image__filename"><?=$file['name']?></span>
+            <?php if ($file['type']=='image'): ?>
+                <div class="demo-card-image mdl-card mdl-shadow--2dp item" data-type="image" data-src="/files/render/o/<?=$file['path']['name'];?>" style="background: url('/files/render/s/<?=$file['path']['name'];?>') center / cover">
+                    <div class="mdl-card__title mdl-card--expand"></div>
+                    <div class="mdl-card__actions">
+                        <span class="demo-card-image__filename"><?=$file['name']?></span>
 
-                    <a href="/files/o/<?=$file['path']['text'].$file['path']['name'];?>" class="download-icon" download>
-                        <div id="tt3" class="icon material-icons">cloud_download</div>
-                    </a>
+                        <a href="/files/o/<?=$file['path']['text'].$file['path']['name'];?>" class="download-icon" download>
+                            <div id="tt3" class="icon material-icons">cloud_download</div>
+                        </a>
+                    </div>
                 </div>
-            </div>
+            <?php else: ?>
+                <div class="demo-card-image mdl-card mdl-shadow--2dp" style="background: url('/assets/img/default_<?=$file['type']?>.png') center / cover">
+                    <div class="mdl-card__title mdl-card--expand"></div>
+                    <div class="mdl-card__actions">
+                        <span class="demo-card-image__filename"><?=$file['name']?></span>
+
+                        <a href="/files/render/o/<?=$file['path']['name'];?>" class="download-icon" download>
+                            <div id="tt3" class="icon material-icons">cloud_download</div>
+                        </a>
+                    </div>
+                </div>
+            <?php endif; ?>
         <?php endforeach; ?>
     <?php else: ?>
         <table class="mdl-data-table mdl-js-data-table mdl-data-table--selectable mdl-shadow--2dp table-gallery">
@@ -25,27 +38,49 @@
             </thead>
             <tbody>
                 <?php foreach ($files as $file): ?>
-                    <tr>
-                        <td>
-<!--                            <img src="" alt="">-->
-                            <img class="item" data-type="image" data-src="/files/render/o/<?=$file['path']['name'];?>" src="/files/render/s/<?=$file['path']['name'];?>" style="width: 50px;">
-                        </td>
-                        <td class="mdl-data-table__cell--non-numeric">
-                            <div class="mdl-textfield mdl-js-textfield mdl-textfield--floating-label">
-                                <input class="mdl-textfield__input file-name" type="text" value="<?=$file['name']?>" data-id="<?=$file['id']?>">
-                                <label class="mdl-textfield__label" for="name">Название...</label>
-                            </div>
-                        </td>
-                        <td class="actions">
-                            <a href="/files/o/<?=$file['path']['text'].$file['path']['name'];?>" class="icon" download>
-                                <div id="tt3" class="icon material-icons">cloud_download</div>
-                            </a>
-                            &nbsp;&nbsp;&nbsp;
-                            <a href="/files/delete/<?=$file['id']?>" class="icon">
-                                <div id="tt3" class="icon material-icons">delete</div>
-                            </a>
-                        </td>
-                    </tr>
+                    <?php if ($file['type']=='image'): ?>
+                        <tr>
+                            <td>
+                                <img class="item" data-type="image" data-src="/files/render/o/<?=$file['path']['name'];?>" src="/files/render/s/<?=$file['path']['name'];?>" style="width: 50px;">
+                            </td>
+                            <td class="mdl-data-table__cell--non-numeric">
+                                <div class="mdl-textfield mdl-js-textfield mdl-textfield--floating-label">
+                                    <input class="mdl-textfield__input file-name" type="text" value="<?=$file['name']?>" data-id="<?=$file['id']?>">
+                                    <label class="mdl-textfield__label" for="name">Название...</label>
+                                </div>
+                            </td>
+                            <td class="actions">
+                                <a href="/files/render/o/<?=$file['path']['name'];?>" class="icon" download>
+                                    <div id="tt3" class="icon material-icons">cloud_download</div>
+                                </a>
+                                &nbsp;&nbsp;&nbsp;
+                                <a href="/files/delete/<?=$file['id']?>" class="icon">
+                                    <div id="tt3" class="icon material-icons">delete</div>
+                                </a>
+                            </td>
+                        </tr>
+                    <?php else: ?>
+                        <tr>
+                            <td>
+                                <img src="/assets/img/default_<?=$file['type']?>.png" style="width: 50px;">
+                            </td>
+                            <td class="mdl-data-table__cell--non-numeric">
+                                <div class="mdl-textfield mdl-js-textfield mdl-textfield--floating-label">
+                                    <input class="mdl-textfield__input file-name" type="text" value="<?=$file['name']?>" data-id="<?=$file['id']?>">
+                                    <label class="mdl-textfield__label" for="name">Название...</label>
+                                </div>
+                            </td>
+                            <td class="actions">
+                                <a href="/files/render/o/<?=$file['path']['name'];?>" class="icon" download>
+                                    <div id="tt3" class="icon material-icons">cloud_download</div>
+                                </a>
+                                &nbsp;&nbsp;&nbsp;
+                                <a href="/files/delete/<?=$file['id']?>" class="icon">
+                                    <div id="tt3" class="icon material-icons">delete</div>
+                                </a>
+                            </td>
+                        </tr>
+                    <?php endif; ?>
                 <?php endforeach; ?>
             </tbody>
         </table>
